Remove TaskData
