+++ src/main/java/hwr/oop/todo/Task.java
@@ -1,19 +1,41 @@
 package hwr.oop.todo;
 
-public class Task extends TaskData{
-    private final String id;
+public class Task {
+    private String title;
+    private String description;
+    private TaskState state;
 
-    public Task(TaskData data, String id){
-        super(data.getTitle(), data.getDescription());
-        this.id = id;
+    public Task(String title, String description){
+        this.title = title;
+        this.description = description;
+        this.state = TaskState.OPEN;
     }
 
-    public static Task fromData(TaskData data){
-        // TODO: Generate id
-        return new Task(data, "abc");
+    public Task(String title) {
+        this(title, "");
     }
 
-    public String getId() {
-        return id;
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public TaskState getState() {
+        return state;
+    }
+
+    public void setState(TaskState state) {
+        this.state = state;
     }
 }
