+++ src/main/java/hwr/oop/todo/TaskController.java
@@ -1,14 +1,19 @@
 package hwr.oop.todo;
 
 import java.util.HashMap;
-import java.util.List;
+
+// TODO:
+//  - Controller testen
+//  - Tag
+//  - Project (kann ggf. von Tag erben)
 
 public class TaskController {
     private HashMap<String, Task> tasks;
 
-    public void addTask(TaskData data){
-        Task task = Task.fromData(data);
-        tasks.put(task.getId(), task);
+    public void addTask(Task task){
+        // TODO: Generate uuid
+        String id = "id:xyz";
+        tasks.put(id, task);
     }
 
     public void getTask(Integer id){
