+++ src/main/java/hwr/oop/todo/TaskState.java
@@ -1,5 +1,5 @@
 package hwr.oop.todo;
 
 public enum TaskState {
-    OPEN, DONE, IN_PROGRESS
+    OPEN, IN_PROGRESS, DONE
 }
